rename init_variable to create_variable
