+++ openpathsampling/features/__init__.py
@@ -1,0 +1,4 @@
+import Configuration
+import Coordinates
+import Velocities
+import Momentum
