+++ openpathsampling/storage/mcstep_store.py
@@ -34,11 +34,11 @@ class MCStepStore(ObjectStore):
         super(MCStepStore, self)._init()
 
         # New short-hand definition
-        self.init_variable('change', 'obj.pathmovechanges', chunksizes=(1,))
-        self.init_variable('active', 'obj.samplesets', chunksizes=(1,))
-        self.init_variable('previous', 'obj.samplesets', chunksizes=(1,))
-        self.init_variable('simulation', 'obj.pathsimulators', chunksizes=(1,))
-        self.init_variable('mccycle', 'int', chunksizes=(1,))
+        self.create_variable('change', 'obj.pathmovechanges', chunksizes=(1,))
+        self.create_variable('active', 'obj.samplesets', chunksizes=(1,))
+        self.create_variable('previous', 'obj.samplesets', chunksizes=(1,))
+        self.create_variable('simulation', 'obj.pathsimulators', chunksizes=(1,))
+        self.create_variable('mccycle', 'int', chunksizes=(1,))
 
     def all(self):
         self.cache_all()
