+++ openpathsampling/storage/object_storage.py
@@ -496,12 +496,12 @@ class ObjectStore(object):
         self.storage.createDimension(self.prefix, 0)
 
         if self.has_name:
-            self.init_variable("name", 'str',
+            self.create_variable("name", 'str',
                                description='A name',
                                chunksizes=tuple([10240]))
 
         if self.json:
-            self.init_variable("json", 'json',
+            self.create_variable("json", 'json',
                                description='A json serialized version of the object',
                                chunksizes=tuple([10240]))
 
@@ -512,7 +512,7 @@ class ObjectStore(object):
     # INITIALISATION UTILITY FUNCTIONS
     # ==============================================================================
 
-    def init_variable(self, name, var_type, dimensions=None, **kwargs):
+    def create_variable(self, name, var_type, dimensions=None, **kwargs):
         """
         Create a new variable in the netCDF storage. This is just a helper
         function to structure the code better.
