--- openpathsampling/storage/snapshot_features.py
@@ -1,255 +1,0 @@
-__author__ = 'jan-hendrikprinz'
-
-import simtk.unit as u
-import numpy as np
-
-from openpathsampling import Configuration, Momentum
-from openpathsampling.netcdfplus import ObjectStore
-
-from openpathsampling.tools import units_from_snapshot
-
-
-class SnapshotFeature(object):
-    """
-    Implements a storable feature that includes a list of variables to be stored.
-    """
-    _variables = []
-
-    @staticmethod
-    def _init(self):
-        pass
-
-class Coordinates(SnapshotFeature):
-    """
-    Feature that will directly store a numpy array in `.coordinates`
-    """
-    _variables = ['coordinates']
-
-    @staticmethod
-    def _init(self):
-        self.init_variable('coordinates', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="coordinate of atom '{ix[1]}' in dimension " +
-                                       "'{ix[2]}' of configuration '{ix[0]}'.",
-                           chunksizes=(1, 'atom', 'spatial')
-                           )
-
-class Velocities(SnapshotFeature):
-    """
-    Feature that will directly store a numpy array in `.velocities`
-    """
-    _variables = ['velocities']
-
-    @staticmethod
-    def _init(self):
-
-        self.init_variable('velocities', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="the velocity of atom 'atom' in dimension " +
-                                       "'coordinate' of momentum 'momentum'.",
-                           chunksizes=(1, 'atom', 'spatial')
-                           )
-
-class Configurations(SnapshotFeature):
-    """
-    Feature that will store a Configuration object in `.configuration`
-    """
-    _variables = ['configuration']
-
-    @staticmethod
-    def _init(self):
-        self.storage.create_store('configurations', ConfigurationStore())
-
-        self.init_variable('configuration', 'lazyobj.configurations',
-                           description="the snapshot index (0..n_configuration-1) of snapshot '{idx}'.",
-                           chunksizes=(1,)
-                           )
-
-class Momenta(SnapshotFeature):
-    _variables = ['momentum']
-
-    @staticmethod
-    def _init(self):
-        self.storage.create_store('momenta', MomentumStore())
-
-        self.init_variable('momentum', 'lazyobj.momenta',
-                           description="the snapshot index (0..n_momentum-1) 'frame' of snapshot '{idx}'.",
-                           chunksizes=(1,)
-                           )
-
-class MomentumStore(ObjectStore):
-    """
-    An ObjectStore for Momenta. Allows to store Momentum() instances in a netcdf file.
-    """
-
-    def __init__(self):
-        super(MomentumStore, self).__init__(Momentum, json=False)
-
-    def _save(self, momentum, idx):
-        self.vars['velocities'][idx, :, :] = momentum.velocities
-
-        if momentum.kinetic_energy is not None:
-            self.vars['kinetic_energy'][idx] = momentum.kinetic_energy
-
-    def _load(self, idx):
-        velocities = self.vars['velocities'][idx]
-        kinetic_energy = self.vars['kinetic_energy'][idx]
-
-        momentum = Momentum(velocities=velocities, kinetic_energy=kinetic_energy)
-
-        return momentum
-
-    def velocities_as_numpy(self, frame_indices=None, atom_indices=None):
-        """
-        Return a block of stored velocities in the database as a numpy array.
-
-        Parameters
-        ----------
-        frame_indices : list of int or None
-            the indices of Momentum objects to be retrieved from the database.
-            If `None` is specified then all indices are returned!
-        atom_indices : list of int of None
-            if not None only the specified atom_indices are returned. Might
-            speed up reading a lot.
-        """
-
-        if frame_indices is None:
-            frame_indices = slice(None)
-
-        if atom_indices is None:
-            atom_indices = slice(None)
-
-        return self.variables['velocities'][frame_indices, atom_indices, :].astype(np.float32).copy()
-
-    def velocities_as_array(self, frame_indices=None, atom_indices=None):
-        """
-        Returns a numpy array consisting of all velocities at the given indices
-
-        Parameters
-        ----------
-        frame_indices : list of int
-            momenta indices to be loaded
-        atom_indices : list of int
-            selects only the atoms to be returned. If None (Default) all atoms
-            will be selected
-
-
-        Returns
-        -------
-        numpy.ndarray, shape = (l,n)
-            returns an array with `l` the number of frames and `n` the number
-            of atoms
-        """
-
-        return self.velocities_as_numpy(frame_indices, atom_indices)
-
-    def _init(self):
-        """
-        Initializes the associated storage to index momentums in it
-        """
-
-        super(MomentumStore, self)._init()
-
-        n_atoms = self.storage.n_atoms
-        n_spatial = self.storage.n_spatial
-
-        units = units_from_snapshot(self.storage._template)
-
-        self.init_variable('velocities', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="the velocity of atom 'atom' in dimension " +
-                                       "'coordinate' of momentum 'momentum'.",
-                           chunksizes=(1, n_atoms, n_spatial),
-                           simtk_unit=units['velocity']
-                           )
-
-        self.init_variable('kinetic_energy', 'float',
-                           chunksizes=(1,),
-                           simtk_unit=units['energy']
-                           )
-
-
-class ConfigurationStore(ObjectStore):
-    """
-    An ObjectStore for Configuration. Allows to store Configuration() instances in a netcdf file.
-    """
-    def __init__(self):
-        super(ConfigurationStore, self).__init__(Configuration, json=False)
-
-    def _save(self, configuration, idx):
-        # Store configuration.
-        self.vars['coordinates'][idx] = configuration.coordinates
-
-        if configuration.potential_energy is not None:
-            self.vars['potential_energy'][idx] = configuration.potential_energy
-
-        if configuration.box_vectors is not None:
-            self.vars['box_vectors'][idx] = configuration.box_vectors
-
-    def get(self, indices):
-        return [self.load(idx) for idx in indices]
-
-    def _load(self, idx):
-        coordinates = self.vars["coordinates"][idx]
-        box_vectors = self.vars["box_vectors"][idx]
-        potential_energy = self.vars["potential_energy"][idx]
-
-        configuration = Configuration(coordinates=coordinates, box_vectors=box_vectors,
-                                      potential_energy=potential_energy)
-        configuration.topology = self.storage.topology
-
-        return configuration
-
-    def coordinates_as_numpy(self, frame_indices=None, atom_indices=None):
-        """
-        Return the atom coordinates in the storage for given frame indices
-        and atoms
-
-        Parameters
-        ----------
-        frame_indices : list of int or None
-            the frame indices to be included. If None all frames are returned
-        atom_indices : list of int or None
-            the atom indices to be included. If None all atoms are returned
-
-        Returns
-        -------
-        numpy.array, shape=(n_frames, n_atoms)
-            the array of atom coordinates in a float32 numpy array
-
-        """
-        if frame_indices is None:
-            frame_indices = slice(None)
-
-        if atom_indices is None:
-            atom_indices = slice(None)
-
-        return self.storage.variables[self.prefix + '_coordinates'][frame_indices, atom_indices, :].astype(
-            np.float32).copy()
-
-    def _init(self):
-        super(ConfigurationStore, self)._init()
-        n_atoms = self.storage.n_atoms
-        n_spatial = self.storage.n_spatial
-
-        units = units_from_snapshot(self.storage._template)
-
-        self.init_variable('coordinates', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="coordinate of atom '{ix[1]}' in dimension " +
-                                       "'{ix[2]}' of configuration '{ix[0]}'.",
-                           chunksizes=(1, n_atoms, n_spatial),
-                           simtk_unit=units['length']
-                           )
-
-        self.init_variable('box_vectors', 'numpy.float32',
-                           dimensions=('spatial', 'spatial'),
-                           chunksizes=(1, n_spatial, n_spatial),
-                           simtk_unit=units['length']
-                           )
-
-        self.init_variable('potential_energy', 'float',
-                           chunksizes=(1,),
-                           simtk_unit=units['energy']
-                           )
-
